src/neuroglancer/requirements.txt
sockjs-tornado==1.0.3
tornado==4.4.2